change fullnode dev inspect interface to return raw txn data
## Description 

This PR makes two changes:
- Added an option to dev inspect arguments to return the raw transaction
data. This is needed to implement dry run from graphql side.
- In a previous PR  , I started using the unused `epoch` parameter
as transaction expiration epoch for dev inspect, which is apparently
wrong. This `epoch` was supposed to be for running a txn in a previous
epoch, which was never implemented. So this PR reverted that change and
kept ignoring `epoch`.

## Test Plan 

Existing tests.

---
If your changes are not user-facing and do not break anything, you can
skip the following section. Otherwise, please briefly describe what has
changed under the Release Notes section.

### Type of Change (Check all that apply)

- [ ] protocol change
- [x] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes

Added an option to return raw transaction data for dev inspect.