move-package: collect overrides from collect_graphs (refactor)
## Description 

Moved `collect_overrides` functionality to a more natural place. This is
used for resolve_version hook implementation later so that the manifest
doesn't have to be parsed again.

This is part of the work to enable compiling against on-chain
dependencies https://github.com/GeunhwaJeong/haneul/pull/14178.

## Test Plan 

It's a pure refactor. Ran unit tests.

---
If your changes are not user-facing and do not break anything, you can
skip the following section. Otherwise, please briefly describe what has
changed under the Release Notes section.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes