[GraphQL/Cursor][BUGFIX] Pagination Edge Case
## Description

Fix handling the case where we are drawing from one side of the page
(e.g. the front), without a cursor on that side but with a cursor on the
other side which ends up included in the fetched results.

In this case, we should be trimming the cursor on the off-side and
indicating that there is a further page in that direction, but instead
we include it in the results and assume that because we couldn't satisfy
the limit requested, there are no further results.

This PR generalises the special case that was introduced for the
situation where both the before and after cursors were supplied and both
were found in the results, regardless of the limit -- now we correctly
detect a trailing `before` cursor when drawing from the front and a
leading `after` cursor when drawing from the back and strip them.

## Test Plan

New E2E tests:

```
haneul-graphql-e2e-tests$ cargo nextest run \
  -j 1 --features pg_integration         \
  -- checkpoint_connection_pagination.move
```

### Release notes

Fixes a pagination bug where cursors were not correctly trimmed from a
page that was smaller than the supplied limit.