[graphql] Exclude health endpoint from logging
## Description 

Due to the way things are set up, the health endpoint is filling up the
logs, so we want to not log it. Unfortunately, all the `cost` logs will
still push an `info` with
`INFO haneul_graphql_rpc::data::pg::query_cost: Estimated cost cost=0.54
max_db_query_cost=20000 exceeds=false`. Adding a Uuid to the log
function in the `cost` module is not trivial. Need to assess if this is
the right way.

Should be merged after 

## Test Plan 

How did you test the new or updated feature?

---
If your changes are not user-facing and do not break anything, you can
skip the following section. Otherwise, please briefly describe what has
changed under the Release Notes section.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes