move toolchain: add testnet binary download fallback
## Description 

See title.

## Test Plan 

Testnet manually locally by changing the required version. It's a bit
tricky to add a test that simulates this. Though, maybe we can pass in
the url to download through a callback and test that way, the change
might just be a little invasive.

---
If your changes are not user-facing and do not break anything, you can
skip the following section. Otherwise, please briefly describe what has
changed under the Release Notes section.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes