[cli] Add client faucet command to get gas
## Description 

Added a new command `haneul client faucet` to make it easier to get gas
from `devnet`, `testnet`, or `local` network. If a different network
than the usual ones (`fullnode.network.haneul.io`) is used, it will
complain and ask to pass the url to its faucet.

There are two options for the command:
`--url`, which is the URL to the gas faucet server
`--address`, which can be either an address or the alias name
corresponding to an address in the wallet.

### Usage
```
haneul client faucet => will use the active network and active address
haneul client faucet --address alias_name => will use the address corresponding to this alias, and the active network
haneul client faucet --url => will use the active address and the faucet server at that endpoint
haneul client faucet --address alias_name --url https://myfaucet.myhaneul.io/v5/gas => will use that address corresponding to the alias, and the faucet server at that endpoint.
```

### Output
It will only output that the request was successful and the task id that
it got from the faucet server. The user has to manually check for the
gas coin if it was sent to the address or not (due to rate-limits, an
automated approach fails easily).

## Test Plan 

<img width="1253" alt="image"
src="https://github.com/GeunhwaJeong/haneul/assets/135084671/584746b8-50aa-483a-b78c-285f8377c1ad">

---
If your changes are not user-facing and do not break anything, you can
skip the following section. Otherwise, please briefly describe what has
changed under the Release Notes section.

### Type of Change (Check all that apply)

- [ ] protocol change
- [x] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes
Added a new command `haneul client faucet` to make it easier to get gas
from `devnet`, `testnet`, or a `local` network. If a different network
than the usual ones (`fullnode.network.haneul.io`) is used, it will
complain and ask to pass the URL to the faucet server.

There are two options for the command:
`--url`, which is the URL to the gas faucet server
`--address`, which can be either an address or the alias name
corresponding to an address in the wallet.

### Usage
```
haneul client faucet => will use the active network and active address
haneul client faucet --address alias_name => will use the address corresponding to this alias, and the active network
haneul client faucet --url => will use the active address and the faucet server at that endpoint
haneul client faucet --address alias_name --url https://myfaucet.myhaneul.io/v5/gas => will use that address corresponding to the alias, and the faucet server at that endpoint.
```

### Output
It will only output that the request was successful and instruct the
user to wait up to 60 seconds and then manually check (`haneul client gas`)
for the gas coin if it was sent to the address or not.