[Consensus 2.0] introduce block signer
## Description 

Adding the block signer using the message `intents`. A new app is also
introduced for the intents with name `Consensus` and the scope
`BlockDigest`. For the signature verification for now I've just
introduced a method `verify` in `SignedBlock` to have a way to verify
the signatures . Refactoring this part could be done on follow up PRs
once we deal with the block verification anyways.

## Test Plan 

CI

---
If your changes are not user-facing and do not break anything, you can
skip the following section. Otherwise, please briefly describe what has
changed under the Release Notes section.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes