[GraphQL/Docs] Tweaking "Getting Started"
## Description

Noticed some small things that I thought would be helpful to change
while preparing the docs for cursors and pagination:

- Name the "Getting Started" guide as "Querying Haneul GraphQL RPC" instead
of just "Haneul GraphQL RPC" so that its name includes a verb, to match
other guides.
- Remove the reference to Github's GraphQL docs here, because that is
already present in the conceptual docs (which we link to from here).
- Flatten out the "Examples" section, as that is mostly what this page
is about.
- Standardise how we present variables across examples.
- Standardise indent level in examples (2 spaces).
- Use the present participle (progressive tense) for all the example
titles ("Finding" vs "Find"). We were using both tenses before -- I
don't feel strongly about one or the other but it's good to standardise.
- Remove reference to tracking `UpgradeCap`s using the `objectModified`
query -- because this is not doable in all cases (e.g. when the
`UpgradeCap` is wrapped).
- Remove reference to `edges { cursor }` in the "Filtering transaction
blocks by a function" example, because we wouldn't typically use cursors
that way.

I also believe the "Limits" section belongs in its own page, in the
"Concepts" section, but I will handle that in a follow-up PR.

## Test Plan

:eyes:

Advice appreciated on how to build and view the docs locally for changes
like this, but I'll otherwise be relying on Vercel.