[graphql] Improve the health check
## Description 

Improve the health check by removing the query to the DB and replacing
it with a status code 200 if all is OK, or 500 on error. It also cleans
up the previous health check checks in the logging logic.

It also moves metrics and connection under a `AppState` struct as `axum`
does not allow states with different types on the router, so we need to
bundle them under one state and then have `FromRef` implementations for
the substates.

## Test Plan 

Existing tests, manually analyzing the logs.

---
If your changes are not user-facing and do not break anything, you can
skip the following section. Otherwise, please briefly describe what has
changed under the Release Notes section.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes