[easy][haneul-package-resolver] Switch package resolver tests to using insta test for expected value tests
## Description 

Converts the expected value tests in the haneul-package-resolver crate to
using insta test. This makes it easier to update/review and also to have
larger resolution tests.

Note that I kept alternate display formatting for the layouts since that
makes things easier to read in the snapshots.

## Test Plan 

Make sure existing tests pass, and that the new expected insta snapshots
match the old ones.