[consensus] Add universal committer
## Description
The `UniversalCommitter` uses a collection of `BaseCommitter`'s to
commit a sequence of leaders. It will be configured to enable
multi-leader & pipelining in Mysticeti.

Note: Feel free to ignore test util changes as most of that will be
reviewed and checked in with
[PR](https://github.com/GeunhwaJeong/haneul/pull/15806)

## Test Plan
Added a few basic tests to ensure things are working, but will add more
tests in a follow up PR so this PR can remain small.