[move-compiler] Parsing error recovery for top-level definitions
## Description 

This PR attempts to provide error recovery during parsing for top-level
module definitions.

Conceptually what we want is to stop parsing a top-level definition if a
parsing error has been encountered, skip this top-level definition
altogether, and move on to parsing the next one.

The challenge is in the "skipping" part - how do we "reset" parser to
the right state so that it can start parsing the next top-level
definition.

The idea is as follows. When encountering a parsing error when parsing
any top-level definition, we keep advancing the parser until we see the
beginning of the next definition (e.g., `#` starting attributes list,
`fun` keyword, `const` keyword, etc.). Then we resume parsing and
eventually return a partially complete parsed AST.

One wrinkle here is what happens when more than one module is defined in
the same file, and the definition of the previous module ends in a
parsing error:
```
module ParseError::M3 {
    const c: u64 = 7;

    const d
}

#[test]
module ParseError::M4 {
    const c: u64 = 7;
}
```
Upon encountering parsing error at the end of `M3` we would keep
advancing the parser searching of the next top-level definition in `M3`.
This would not be a problem if `M3` was the only module defined in a
file as the parser would eventually encounter EOF and quit.

As it happens, it's not a big problem even if `M3` is succeeded by `M4.
The parser will encounter start of attributes definition for `M4`. It
will "think" it's another top-level definition of `M3` and try to parse
it. It will fail upon encountering the `module` keyword, but it will be
in the correct place to simply finish parsing `M3` and start parsing
`M4` (with optional attributes that have to be kept on the side for
this).

## Test Plan 

Tests have been updated and new tests added.

---
If your changes are not user-facing and do not break anything, you can
skip the following section. Otherwise, please briefly describe what has
changed under the Release Notes section.

### Type of Change (Check all that apply)

- [ ] protocol change
- [x] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes
Developers may see more compiler diagnostics than before as parsing
errors no longer stop compilation and diagnostics from later compilation
stages my get included as well in the compilation result.