Update base64 handling
## Description 

This updates the base64 code in the bcs library to use the `atob` and
`btoa` built-ins, which should be faster, and are available in nearly
every JS environment. In places these aren't available, they're trivial
to polyfill, and can be part of our standard list of globals we expect
folks to polyfill (`fetch`, `TextEncoder`, `crypto`, etc.)

## Test Plan 

How did you test the new or updated feature?

---
If your changes are not user-facing and do not break anything, you can
skip the following section. Otherwise, please briefly describe what has
changed under the Release Notes section.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes