Version Packages
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and publish to npm
yourself or [setup this action to publish
automatically](https://github.com/changesets/action#with-publishing). If
you're not ready to do a release yet, that's fine, whenever you add more
changesets to main, this PR will be updated.


# Releases
## haneullabs/create-dapp@0.2.3

### Patch Changes

-   haneullabs/dapp-kit@0.12.1

## haneullabs/dapp-kit@0.12.1

### Patch Changes

-   Updated dependencies [59339b7289]
    -   haneullabs/zksend@0.2.1

## haneullabs/zksend@0.2.1

### Patch Changes

-   59339b7289: fix listing assets for empty links