[Consensus 2.0] propagate errors from Core to upstream
## Description 

Propagates the errors from Core to upstream components (namely
CoreThread) to handle. When the error is a `Shutdown` one then we close
the core loop and exit, otherwise we panic as it is deemed a fatal one.
Since we are doing this I've also propagated an error from the commit
observer to follow a uniform approach.

## Test Plan 

CI

---
If your changes are not user-facing and do not break anything, you can
skip the following section. Otherwise, please briefly describe what has
changed under the Release Notes section.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes