Fix dir prefix that is used for backfills
## Description 

The dir prefix path is today not acknowledged during the setup phase
where pipeline starts from a particular checkpoint. This will cause the
pipeline to re-run from the same checkpoint when doing backfills. This
PR fixes the particular issue.

## Test Plan 

Ran locally, verified the path used during re-run is the backfill path
(not the main path)