Version Packages
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and publish to npm
yourself or [setup this action to publish
automatically](https://github.com/changesets/action#with-publishing). If
you're not ready to do a release yet, that's fine, whenever you add more
changesets to main, this PR will be updated.


# Releases
## haneullabs/enoki@0.2.0

### Minor Changes

-   7b5132de52: Add support for secret key authentication

## haneullabs/zksend@0.3.0

### Minor Changes

-   7b5132de52: Add support for contract based links

## haneullabs/create-dapp@0.2.6

### Patch Changes

-   haneullabs/dapp-kit@0.12.4

## haneullabs/dapp-kit@0.12.4

### Patch Changes

-   Updated dependencies [7b5132de52]
    -   haneullabs/zksend@0.3.0