[graphql] Fix metrics issue with elapsed time not being correct
## Description 

Fix metrics issue with elapsed time not being correctly converted to
seconds. It is now standardized everywhere that any elapsed argument
should be of type `Duration`, and on that variable we call
`as_secs_f64()` to pass it to the prometheus metric.
This should fix the problem with the latency metric in our dashboards.

## Test Plan 

👀 

---
If your changes are not user-facing and do not break anything, you can
skip the following section. Otherwise, please briefly describe what has
changed under the Release Notes section.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes