move lock: in-place dependency graph updates
## Description 

This PR makes it so that `write_to_lock` in dependency graph code only
populates the `Move.lock` with the relevant state it ought to, leaving
everything else in the lock file alone. This is in contrast to the
current "wipe everything, populate dependencies, and expect orthogonal
information to be populated afterwards on every invocation". This
applies to the following TOML values in `Move.lock`:

- `manifest_digest`
- `deps_digest`
- `dependencies` (for root package)
- `dev-dependencies` (for root package)
- `[[move.packages]]` array of table

## Test Plan 

Updated tests. Lock file re-generation may remove empty lines in
`Move.lock` and is a user-visible change, though not semantically
meaningful.

---
If your changes are not user-facing and do not break anything, you can
skip the following section. Otherwise, please briefly describe what has
changed under the Release Notes section.

### Type of Change (Check all that apply)

- [ ] protocol change
- [x] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes