[move] Add error constants to source language (duplicate of )
## Description 

NB: This is a duplicate of the already-approved  since those
changes were accidentally landed into the lower commit.

This PR adds support for "clever assertions" to the Move source language
for Move 2024 (NB: no changes to the bytecode format!).

Clever assertions allow you to use non-u64 consts as abort codes, as
long as they are annotated with the new `#[error]` attribute:

```move
module std::AModule {
    #[error]
    const EIsThree: vector<u8> = b"EIsThree";

    public fun double_except_three(x: u64): u64 {
        assert!(x != 3, EIsThree);
        x * x
    }
}
```

Note that outside of abort context, constants annotated `#[error]` will
still be typechecked normally, e.g., the following is a valid program in
this setting
```move
#[error]
const ENotFound: bool = true;

fun f() {
    // usage in arg1 is as a bool and in arg2 as a u64
    assert!(ENotFound, ENotFound);
}
```

## Assertions with No Abort Codes

This also adds support for assertions with no abort codes:
```move
fun cool(x: u64) {
  assert!(x == 42);
}
```

This will be compiled to a tagged `u64` abort code, constant identifier,
and constant value fields set to `0xFFFF`, and `0xFFFF` (see below).

## Compilation scheme

Non-u64 constants used in an abort context, and that have been annotated
with `#[error(code = <n>)]` will turn into a tagged abort code (a `u64`)
with the following layout
```
    // |<tagbit>|<reserved>|<line number>|<identifier index>|<constant index>|
    //   1-bit    15-bits      16-bits        16-bits          16-bits
``` 

i.e., we tag these special abort codes, and then bit-pack source line
number, then the index in the constant pool that holds the string name
of the constant that was used, and then the index in the constant pool
of the actual constant value. Note, that in some cases the index of the
constant name into the constant pool and the index into the constant
pool for the constants value may be the same, e.g., if you have
something like
```move
#[error]
const MyError: vector<u8> = b"MyError";
```

In the case of "naked assertions" with no abort code, we simply set the
tag bit, and line number parts of the bitfield, and set the error code,
identifier, and constant value indices to their max value (`0xFFFF`). If
the line number is greater than `u16::MAX` then the value is clamped at
that value.

## Test Plan 

Added additional tests for these behaviors. 

---
If your changes are not user-facing and do not break anything, you can
skip the following section. Otherwise, please briefly describe what has
changed under the Release Notes section.

### Type of Change (Check all that apply)

- [ ] protocol change
- [X] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes

## Description 

Describe the changes or additions included in this PR.

## Test Plan 

How did you test the new or updated feature?

---
If your changes are not user-facing and do not break anything, you can
skip the following section. Otherwise, please briefly describe what has
changed under the Release Notes section.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes