Version Packages
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and publish to npm
yourself or [setup this action to publish
automatically](https://github.com/changesets/action#with-publishing). If
you're not ready to do a release yet, that's fine, whenever you add more
changesets to main, this PR will be updated.


# Releases
## haneullabs/zksend@0.4.0

### Minor Changes

- c481dec663: removed listClaimableAssets, and added new assets and
claimed properties to link instances
-   c481dec663: Use contract by default for new links
-   c481dec663: Add helper for bulk link creation
-   c481dec663: Removed options for filtering claims
-   c481dec663: renamed loadOwnedData to loadAssets

## haneullabs/create-dapp@0.2.8

### Patch Changes

-   haneullabs/dapp-kit@0.12.6

## haneullabs/dapp-kit@0.12.6

### Patch Changes

-   Updated dependencies [c481dec663]
-   Updated dependencies [c481dec663]
-   Updated dependencies [c481dec663]
-   Updated dependencies [c481dec663]
-   Updated dependencies [c481dec663]
    -   haneullabs/zksend@0.4.0