[cli][easy] Add serialize-signed-transaction and serialize-unsigned transaction flags for haneul client ptb command
## Description 

Add support for the `serialize-signed-transaction` and
`serialize-unsigned-transaction` flags that output transaction bytes
instead of executing the transaction for the `haneul client ptb` command.

## Test Plan 

Existing tests as this uses the `serialize-or-execute` macro.

---
If your changes are not user-facing and do not break anything, you can
skip the following section. Otherwise, please briefly describe what has
changed under the Release Notes section.

### Type of Change (Check all that apply)

- [ ] protocol change
- [x] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes
Added support for the `serialize-signed-transaction` and
`serialize-unsigned-transaction` flags that output transaction bytes
instead of executing the transaction for the `haneul client ptb` command.