[GraphQL] [indexer] split type field into different components
For event and object types stored in the indexer tables, this PR stores
the `package`, `module` and `name` of the `StructTag` in addition to the
full type information. With this change all the queries we do to filter
by type in graphql will be exact match instead of using `LIKE`.

I chose to go with still keeping the full type field for backward
compatibility (since indexer and gql are on different release
schedules), testability (so we can easily test performance difference
using the old and new way against the same db) and simplicity (so we
don't need to worry too much about type params).

Existing graphql tests and added indexer ingestion tests.

---
If your changes are not user-facing and do not break anything, you can
skip the following section. Otherwise, please briefly describe what has
changed under the Release Notes section.

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration