Version Packages
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and publish to npm
yourself or [setup this action to publish
automatically](https://github.com/changesets/action#with-publishing). If
you're not ready to do a release yet, that's fine, whenever you add more
changesets to main, this PR will be updated.


# Releases
## haneullabs/create-dapp@0.2.11

### Patch Changes

-   haneullabs/dapp-kit@0.12.9

## haneullabs/dapp-kit@0.12.9

### Patch Changes

-   Updated dependencies [381777dd4a]
-   Updated dependencies [c31e86c534]
    -   haneullabs/zksend@0.4.3

## haneullabs/zksend@0.4.3

### Patch Changes

-   381777dd4a: Make contract optional in ZkLSendLinkBuilder.createLinks
- c31e86c534: Handle the fetch function for environments that don't
support window