[move] Fix vfs relativization
## Description 

- pathdiff did not seem to handle windows paths very well, at least in
conjunction with the way VFS was converting them to a string
- Should fix 

## Test Plan 

- Built on windows

---
If your changes are not user-facing and do not break anything, you can
skip the following section. Otherwise, please briefly describe what has
changed under the Release Notes section.

### Type of Change (Check all that apply)

- [ ] protocol change
- [X] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes

Fixed issues with `haneul move` not working in `cmd.exe` on Windows.