Stefan/win support
## Description 

On Windows we have reports (+ repro locally) that `cargo build` fails at
`haneulop-cli` boilerplate code. In the repo, there is a file, so
presumably, when the repo is cloned, the file is copied as well. The
`include_dir!` macro expects a directory, but in the end it is a file,
which leads to an error.

after-ephemera would appreciate your feedback on this one such that we
can get it to compile on Windows. Thanks!

## Test Plan 

Existing tests.

---
If your changes are not user-facing and do not break anything, you can
skip the following section. Otherwise, please briefly describe what has
changed under the Release Notes section.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes