[transfer-to-object] TTO support in test scenario
## Description 

This adds support for transfer-to-object/Receiving tickets in the test
scenario.

In order to do this a "proper" object store needed to be added to the
test scenario since we cannot count on caching to keep the object for us
like we can with dynamic fields in test scenario tests.

This adds such a test by creating a shared `InMemoryStorage` (wrapped in
a mutex since tests can be multi-threaded and we need internal
mutability). Two handles are then created for this -- one for the
storage read operations, and another is added to the context extension
so that we can place "objects-to-be-received" into this store when we
create the `Receiving` ticket.

This might generally be useful for other tests as well/refactoring the
test scenario to be more realistic but I will leave any such
investigations for another day.

## Test Plan 

Added tests for the new features to the test scenario tests.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes