[nre] update config template for consensus 2.0
## Description 

In Consensus 2.0, we are looking into using TCP for communications. So
making sure TCP ports are open in the sample configurations. Also update
the logging setting.

## Test plan 

n/a

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: