[cli] Add dry run for client commands
## Description 

This PR adds the `dry-run` flag for the following `haneul client` commands:
- call
- merge-coin
- split-coin
- pay
- pay-haneul
- pay-all-haneul
- transfer
- transfer-haneul
- publish
- upgrade
- ptb

Note that the `gas-budget` is still required to be passed despite using
`dry-run`. That is intentional and it will be changed to become optional
in a subsequent PR that adds automatic gas budget estimation.

## Test Plan 

Existing tests and added a several new tests, including one that checks if the dry run is successful and if the given gas object for an execution is the one that
was actually used - this is needed because the changes touch some code
related to selecting the gas object.
```
cargo nextest run 
cargo test -- test_dry_run test_pay test_transfer
```

---
If your changes are not user-facing and do not break anything, you can
skip the following section. Otherwise, please briefly describe what has
changed under the Release Notes section.

### Type of Change (Check all that apply)

- [ ] protocol change
- [x] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes
Added the `--dry-run` flag for the following `haneul client` commands to execute a dry run of the transaction instead of executing it and storing its effects on chain. Using `dry-run` you can also get the estimated gas budget needed to execute that transaction. The commands which support `--dry-run` flag are:
- call
- merge-coin
- split-coin
- pay
- pay-haneul
- pay-all-haneul
- transfer
- transfer-haneul
- publish
- upgrade
- ptb