detect loss of liveness in tests
In simtests, and also if the `PANIC_ON_NEW_CHECKPOINT_TIMEOUT` env var
is set, we panic if we do not receive a certified checkpoint for too
long. This condition should encompass more or less everything we mean
when we say "loss of liveness".

In order to get tests to pass under this stricter condition, several
fixes were required:
- changes to the crash failpoints to prevent crashes from happening too
close together in time
- disable pruning in crash tests

Additionally, some tests cannot pass in this condition, so we suppress
the panic on a per-test basis.