[TrafficController] Delegate error weighting to the server
## Description 

The list of servers that need to run traffic controller is 2 and will be
3 with graphql, perhaps more later.

Instead of requiring that traffic controller understand different error
types for each server, allow the server to define the importance of an
error type by replacing `ServiceResponse` in `TrafficTally` with a
`Weight`, which is a float between 0 and 1, generated by each server.
For errors that the server does not care about, weight is set to 0, and
ignored by traffic controller during error tallying. For spam tallying,
we disregard weight and consider all tallies.

## Test plan 

This logic currently only serves to simplify some error case statements.
In later PR's, we will test this logic against error-aware policies
(there are currently none). So for now, just ensure all existing tests
pass

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: