Version Packages
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and publish to npm
yourself or [setup this action to publish
automatically](https://github.com/changesets/action#with-publishing). If
you're not ready to do a release yet, that's fine, whenever you add more
changesets to main, this PR will be updated.


# Releases
## haneullabs/zklogin@0.6.0

### Minor Changes

-   43c2424809: add support for verify zklogin signature in sdk

## haneullabs/enoki@0.2.4

### Patch Changes

-   Updated dependencies [43c2424809]
    -   haneullabs/zklogin@0.6.0