[replay] Skip persisting unsupported transactions
## Description 

There are two types of transactions that are currently not supported
today:
- System transactions (we should fix this)
- Transactions with protocol version < 16. This is due to the dependency
to wrapped tombstones, which we have no way of fetching.
When we see unsupported transactions, we should return error instead of
Ok. And in that case we skip persisting them.

## Test plan 

CI

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: